fix: resolve all real clang-tidy findings across entire codebase
- readability-math-missing-parentheses: add explicit parentheses at 39
  sites across 13 files to clarify operator precedence
- performance-for-range-copy / unnecessary-copy-initialization: 14
  range-for loop variables and copy-inits changed to const refs
- bugprone-crtp-constructor-accessibility: 5 CRTP mixin bases in
  strongtype.h given protected default constructors
- bugprone-unhandled-self-assignment: history.h HistMessage::operator=
  now guards against self-assignment
- bugprone-branch-clone: collapsed 6 switch/if duplicate branches in
  iaudiosink.h, filetransferwidget.cpp, toxcoreerrorparser.cpp,
  network.cpp, history.cpp, chatformheader.cpp
- cppcoreguidelines-init-variables / pro-type-member-init: ~180
  uninitialized locals and member variables given zero initializers
  across audio, chatlog, core, ipc, model, net, persistence, platform,
  video, widget
- performance-unnecessary-value-param: ~116 function parameters changed
  from by-value to const-ref for non-trivial types (QString, QByteArray,
  ToxFile, shared_ptr, etc.) with matching header updates
- bugprone-narrowing-conversions: explicit static_cast added at ~266
  sites (qsizetype->int, size_t->ptrdiff_t, qreal->float, etc.)
- bugprone-implicit-widening-of-multiplication-result: pre-cast operands
  before multiply where result is stored in a wider type
- cppcoreguidelines-pro-type-static-cast-downcast: static_cast downcasts
  replaced with qobject_cast (QObject hierarchy) or dynamic_cast
  (QGraphicsItem hierarchy) throughout widget and chatlog code

All 35 ASan/UBSan tests pass.

+++ src/model/chathistory.cpp
@@ -219,7 +219,7 @@ void ChatHistory::onFileUpdated(const ToxPk& sender, const ToxFile& file)
             // initializing. If this is changed in the session chat log we'll end up
             // with a different order when loading from history
             history->addNewFileMessage(chat.getPersistentId(), file.resumeFileId, file.fileName,
-                                       file.filePath, file.progress.getFileSize(), sender,
+                                       file.filePath, static_cast<int64_t>(file.progress.getFileSize()), sender,
                                        QDateTime::currentDateTime(), username);
             break;
         }
