+++ src/model/toxclientstandards.h
@@ -9,7 +9,7 @@
 
 namespace ToxClientStandards {
 // From TCS 2.2.4, max valid avatar size is 64KiB
-constexpr static uint64_t MaxAvatarSize = 64 * 1024;
+constexpr static uint64_t MaxAvatarSize = static_cast<uint64_t>(64) * 1024;
 constexpr bool IsValidAvatarSize(uint64_t fileSize)
 {
     return fileSize <= MaxAvatarSize;
